+++ RSPiX/Src/GREEN/BLiT/BLIT.cpp
@@ -156,7 +156,7 @@
 // force instantiation
 template void rspClipPlot<UCHAR>(UCHAR color, RImage* pimDst,int16_t sX,int16_t sY,const RRect* prClip);
 template void rspClipPlot<USHORT>(USHORT color, RImage* pimDst,int16_t sX,int16_t sY,const RRect* prClip);
-template void rspClipPlot<ULONG>(ULONG color, RImage* pimDst,int16_t sX,int16_t sY,const RRect* prClip);
+template void rspClipPlot<uint32_t>(uint32_t color, RImage* pimDst,int16_t sX,int16_t sY,const RRect* prClip);
 
 
 void instantiatePlot(void);
@@ -166,7 +166,7 @@ void instantiatePlot()
 
 	rspPlot((UCHAR)0,pim,(int16_t)0,(int16_t)0);
 	rspPlot((USHORT)0,pim,(int16_t)0,(int16_t)0);
-	rspPlot((ULONG)0,pim,(int16_t)0,(int16_t)0);
+	rspPlot((uint32_t)0,pim,(int16_t)0,(int16_t)0);
 
 	}
 
@@ -313,7 +313,7 @@ template int16_t rspLasso<UCHAR>(UCHAR ignoreColor,RImage* pimSrc,int16_t &x,int
 						int lFlag,int rFlag,int tFlag,int bFlag);
 template int16_t rspLasso<USHORT>(USHORT ignoreColor,RImage* pimSrc,int16_t &x,int16_t &y,int16_t &w,int16_t &h,
 						int lFlag,int rFlag,int tFlag,int bFlag);
-template int16_t rspLasso<ULONG>(ULONG ignoreColor,RImage* pimSrc,int16_t &x,int16_t &y,int16_t &w,int16_t &h,
+template int16_t rspLasso<uint32_t>(uint32_t ignoreColor,RImage* pimSrc,int16_t &x,int16_t &y,int16_t &w,int16_t &h,
 						int lFlag,int rFlag,int tFlag,int bFlag);
 
 
@@ -326,7 +326,7 @@ void instantiateLasso()
 
 	rspLasso( (UCHAR)i, pim, i,i,i,i, 1,1,1,1);
 	rspLasso( (USHORT)i, pim, i,i,i,i, 1,1,1,1);
-	rspLasso( (ULONG)i, pim, i,i,i,i, 1,1,1,1);
+	rspLasso( (uint32_t)i, pim, i,i,i,i, 1,1,1,1);
 
 	}
 
@@ -1141,7 +1141,7 @@ int16_t rspRect(U32 color,RImage* pimDst,int16_t sX,int16_t sY,int16_t sW,int16_
 	if ( (sAlign & 3) == 0)	
 		{
 		// 32-bit copy
-		ULONG ulColor = color;
+		uint32_t ulColor = color;
 		ulColor += (color << 8); // 16-bit
 		ulColor += (ulColor << 16); // 32-bit
